Moved all plugin configs to individual files
