+++ init.lua
@@ -1,7 +1,6 @@
 ---------------------------------- Aliases ------------------------------------
 
 A = {}
-
 A.augroup = vim.api.nvim_create_augroup
 A.autocmd = vim.api.nvim_create_autocmd
 A.del = vim.keymap.del
