+++ lua/plugins.lua
@@ -1,399 +1,22 @@
-local lsp = require('lsp')
-
 return {
-    lsp,
-    ---- Autopairs
-    {
-        'windwp/nvim-autopairs',
-        event = "InsertEnter",
-        opts = {} -- this is equalent to setup({}) function
-    },
-    ---- Comment
-    {
-        'numToStr/Comment.nvim',
-        event = { "BufReadPre", "BufNewFile" },
-        config = function()
-            require('Comment').setup({
-                toggler = { line = '<Leader>`', block = '<Leader>~', },
-                opleader = { line = '<Leader>`', block = '<Leader>~', },
-                mappings = { extra = false, extended = false, },
-            })
-        end
-    },
-    ---- Dashboard
-    {
-        'glepnir/dashboard-nvim',
-        dependencies = { 'nvim-tree/nvim-web-devicons' },
-        config = function()
-            A.map('n', '<Leader><CR>', ':Dashboard<CR>', A.opts)
-            require('dashboard').setup({
-                theme = 'doom',
-                config = {
-                    center = {
-                        { icon = '    ', desc = 'Plugin Manager ', action = 'Lazy' },
-                        { icon = '    ', desc = 'LSP Manager ', action = 'Mason' },
-                        { icon = '    ', desc = 'Config         ', action = 'e ~/.config/nvim/init.lua' },
-                    },
-                    header = {
-                        [[                                                       ]],
-                        [[                                                       ]],
-                        [[                                                       ]],
-                        [[                                                       ]],
-                        [[                                                       ]],
-                        [[                                                       ]],
-                        [[                                                       ]],
-                        [[                                                       ]],
-                        [[                                                       ]],
-                        [[                                                       ]],
-                        [[                                                       ]],
-                        [[                                                       ]],
-                        [[                                                       ]],
-                        [[                                                       ]],
-                        [[ ███╗   ██╗ ███████╗ ██████╗  ██╗   ██╗ ██╗ ███╗   ███╗]],
-                        [[ ████╗  ██║ ██╔════╝██╔═══██╗ ██║   ██║ ██║ ████╗ ████║]],
-                        [[ ██╔██╗ ██║ █████╗  ██║   ██║ ██║   ██║ ██║ ██╔████╔██║]],
-                        [[ ██║╚██╗██║ ██╔══╝  ██║   ██║ ╚██╗ ██╔╝ ██║ ██║╚██╔╝██║]],
-                        [[ ██║ ╚████║ ███████╗╚██████╔╝  ╚████╔╝  ██║ ██║ ╚═╝ ██║]],
-                        [[ ╚═╝  ╚═══╝ ╚══════╝ ╚═════╝    ╚═══╝   ╚═╝ ╚═╝     ╚═╝]],
-                        [[                                                       ]],
-                        [[                                                       ]],
-                        [[                                                       ]],
-                        [[                                                       ]],
-                        [[                                                       ]],
-                        [[                                                       ]]
-                    },
-                },
-            })
-        end
-    },
-    ---- Fugitive
-    {
-        'tpope/vim-fugitive',
-        config = function()
-            vim.api.nvim_create_user_command('Diff', function(input)
-                local branch = input.args ~= "" and input.args or "main"
-                vim.cmd('Gvdiffsplit! ' .. branch)
-            end, { nargs = '?' })
-            vim.cmd [[command! -nargs=0 Blame G blame]]
-            vim.cmd [[command! -nargs=0 Merge G mergetool]]
-        end
-    },
-    ---- Git Signs
-    {
-        'lewis6991/gitsigns.nvim',
-        event = { "BufReadPre", "BufNewFile" },
-        config = function()
-            require('gitsigns').setup({
-                on_attach = function()
-                    local gs = package.loaded.gitsigns
-                    A.map('n', ']c', function()
-                        if vim.wo.diff then return ']c' end
-                        vim.schedule(function() gs.next_hunk() end)
-                        return '<Ignore>'
-                    end, { expr = true })
-                    A.map('n', '[c', function()
-                        if vim.wo.diff then return '[c' end
-                        vim.schedule(function() gs.prev_hunk() end)
-                        return '<Ignore>'
-                    end, { expr = true })
-                    A.map('n', 'gc', gs.preview_hunk)
-                    A.map('n', 'gb', function() gs.blame_line { full = true } end)
-                end
-            })
-        end
-
-    },
-    ---- Indent Blank Line
-    {
-        'lukas-reineke/indent-blankline.nvim',
-        event = { "BufReadPre", "BufNewFile" },
-        main = "ibl",
-        config = function()
-            local hooks = require "ibl.hooks"
-            hooks.register(
-                hooks.type.WHITESPACE,
-                hooks.builtin.hide_first_space_indent_level
-            )
-            require("ibl").setup({
-                indent = { char = '▏' },
-                scope = { enabled = false },
-                exclude = {
-                    filetypes = { 'dashboard', 'haskell', 'help', 'python', 'undotree' },
-                    buftypes = { 'nofile', 'terminal' }
-                }
-            })
-        end
-    },
-    ---- Lualine
-    {
-        'nvim-lualine/lualine.nvim',
-        dependencies = { 'nvim-tree/nvim-web-devicons' },
-        event = { "BufReadPre", "BufNewFile" },
-        config = function()
-            local big_screen = function() return vim.fn.winwidth(0) > 90 end
-
-            local lsp_status = function()
-                local clients = vim.lsp.get_active_clients({ bufnr = 0 })
-                if next(clients) == nil then return "No LSP Connected" end
-
-                local client_names = {}
-                for _, client in ipairs(clients) do
-                    table.insert(client_names, client.name)
-                end
-
-                return "LSP: " .. table.concat(client_names, ", ")
-            end
-
-            require 'lualine'.setup({
-                extensions = { 'quickfix' },
-                options = { disabled_filetypes = { 'dashboard', 'NvimTree', 'undotree' } },
-                sections = {
-                    lualine_a = { { 'mode', fmt = function(str) return str:sub(1, 1) end } },
-                    lualine_b = { 'branch', 'diff', 'diagnostics' },
-                    lualine_c = { 'filename' },
-                    lualine_x = { { lsp_status, cond = big_screen } },
-                    lualine_y = { { 'progress', cond = big_screen } },
-                    lualine_z = { { 'location', cond = big_screen } }
-                },
-            })
-        end
-    },
-    ---- Markdown Preview
-    {
-        "iamcco/markdown-preview.nvim",
-        cmd = { "MarkdownPreviewToggle", "MarkdownPreview", "MarkdownPreviewStop" },
-        build = "cd app && yarn install",
-        init = function()
-            vim.g.mkdp_filetypes = { "markdown" }
-        end,
-        ft = { "markdown" },
-    },
-    ---- Nvim Tree
-    {
-        'nvim-tree/nvim-tree.lua',
-        dependencies = { 'nvim-tree/nvim-web-devicons' },
-        keys = { "<C-e>" },
-        config = function()
-            A.map('n', '<C-e>', ':NvimTreeToggle<CR>', A.opts)
-
-            local function nvim_tree_on_attach(bufnr)
-                local api = require('nvim-tree.api')
-
-                local function opts(desc)
-                    return { desc = 'nvim-tree: ' .. desc, buffer = bufnr, noremap = true, silent = true, nowait = true }
-                end
-
-                api.config.mappings.default_on_attach(bufnr)
-
-                A.map('n', 'h', api.node.navigate.parent_close, opts('Close Directory'))
-                A.map('n', 'l', api.node.open.replace_tree_buffer, opts('Open: In Place'))
-                A.map('n', '<BS>', api.tree.change_root_to_parent, opts('Up'))
-                A.map('n', '<C-s>', api.node.open.horizontal, opts('Open: Horizontal Split'))
-                A.map('n', '?', api.tree.toggle_help, opts('Help'))
-
-                A.del('n', '<C-e>', { buffer = bufnr })
-                A.del('n', '<C-x>', { buffer = bufnr })
-            end
-
-            require('nvim-tree').setup({
-                actions             = {
-                    open_file = { quit_on_open = true },
-                },
-                renderer            = {
-                    indent_markers = { enable = true }
-                },
-                respect_buf_cwd     = true,
-                on_attach           = nvim_tree_on_attach,
-                sync_root_with_cwd  = true,
-                update_focused_file = { enable = true, update_root = true },
-                view                = { width = { min = 35 } },
-            })
-        end
-    },
-    ---- Surround
-    {
-        'kylechui/nvim-surround',
-        event = { "BufReadPre", "BufNewFile" },
-        config = function()
-            require("nvim-surround").setup()
-        end
-    },
-    ---- Telescope
-    {
-        'nvim-telescope/telescope.nvim',
-        keys = {"<leader>f", "<leader>g", "<leader>h"},
-        dependencies = {
-            'nvim-lua/plenary.nvim',
-            'nvim-telescope/telescope-fzy-native.nvim',
-            'nvim-telescope/telescope-ui-select.nvim'
-        },
-        config = function()
-            local actions = require('telescope.actions')
-            require('telescope').setup({
-                defaults = {
-                    entry_prefix = "  ",
-                    layout_config = {
-                        horizontal = {
-                            preview_cutoff = 150,
-                            preview_width = 0.45,
-                            prompt_position = "top"
-                        },
-                        width = 0.9
-                    },
-                    mappings = {
-                        i = {
-                            ["<C-j>"] = actions.move_selection_next,
-                            ["<C-k>"] = actions.move_selection_previous,
-                            ["<C-q>"] = actions.smart_send_to_qflist + actions.open_qflist,
-                            ["<C-r>"] = actions.delete_buffer,
-                            ["<C-s>"] = actions.select_horizontal,
-                            ["<C-t>"] = actions.toggle_selection,
-                            ["<M-BS>"] = { "<c-s-w>", type = "command" },
-                        },
-                        n = {
-                            ["<C-t>"] = actions.toggle_selection,
-                        },
-                    },
-                    path_display = { shorten = 5 },
-                    prompt_prefix = " ",
-                    selection_caret = " ",
-                    sorting_strategy = "ascending",
-                },
-            })
-            require('telescope').load_extension('fzy_native')
-            require('telescope').load_extension('ui-select')
-            require('telescope').load_extension('yank_history')
-
-            A.map('n', '<Leader>b', ':Telescope buffers<CR>', A.opts)
-            A.map('n', '<Leader>c', ':Telescope commands<CR>', A.opts)
-            A.map('n', '<Leader>d', ':Telescope help_tags prompt_title=Documentation<CR>', A.opts)
-            A.map('n', '<Leader>f', ':Telescope find_files<CR>', A.opts)
-            A.map('n', '<Leader>g', ':Telescope live_grep<CR>', A.opts)
-            A.map('n', '<Leader>h', ':Telescope oldfiles prompt_title=History<CR>', A.opts)
-            A.map('n', '<Leader>m', ':Telescope keymaps<CR>', A.opts)
-            A.map('n', '<Leader>y', ':Telescope yank_history<CR>', A.opts)
-        end
-    },
-    ---- Tmux Integration
-    {
-        'aserowy/tmux.nvim',
-        config = function()
-            require("tmux").setup({
-                copy_sync = {
-                    redirect_to_clipboard = true,
-                },
-                navigation = {
-                    cycle_navigation = false,
-                    enable_default_keybindings = true,
-                    persist_zoom = true,
-                },
-                resize = {
-                    enable_default_keybindings = true,
-                    resize_step_x = 5,
-                    resize_step_y = 2,
-                }
-            })
-        end
-    },
-    ---- Toggleterm
-    {
-        'akinsho/toggleterm.nvim',
-        version = "*",
-        keys = { "<C-t>" },
-        config = function()
-            require("toggleterm").setup({
-                size = function(term)
-                    if term.direction == "horizontal" then
-                        return 15
-                    elseif term.direction == "vertical" then
-                        return vim.o.columns * 0.5
-                    end
-                end,
-                open_mapping = [[<C-t>]],
-                direction = 'vertical',
-
-            })
-        end
-    },
-    ---- Tokyonight
-    {
-        'folke/tokyonight.nvim',
-        lazy = false,
-        priority = 1000,
-        config = function()
-            require("tokyonight").setup({
-                sidebars = { "qf", "help", "undotree" },
-                lualine_bold = true,
-            })
-            vim.cmd("colorscheme tokyonight")
-        end
-    },
-    ---- Treesitter
-    {
-        'nvim-treesitter/nvim-treesitter',
-        event = { "BufReadPre", "BufNewFile" },
-        build = ':TSUpdate',
-        config = function()
-            require 'nvim-treesitter.configs'.setup({
-                ensure_installed = {
-                    "bash",
-                    "comment",
-                    "css",
-                    "dockerfile",
-                    "gitignore",
-                    "go",
-                    "haskell",
-                    "html",
-                    "javascript",
-                    "json",
-                    "lua",
-                    "markdown",
-                    "python",
-                    "regex",
-                    "rust",
-                    "scss",
-                    "sql",
-                    "toml",
-                    "tsx",
-                    "typescript",
-                    "yaml",
-                },
-                highlight = { enable = true },
-                incremental_selection = {
-                    enable = true,
-                    keymaps = {
-                        init_selection = "+",
-                        node_incremental = "+",
-                        node_decremental = "_",
-                    },
-                },
-                indent = { enable = true }
-            })
-        end
-    },
-    ---- Undo Tree
-    {
-        'mbbill/undotree',
-        event = { "BufReadPre", "BufNewFile" },
-        config = function()
-            A.map('n', '<Leader>u', ':UndotreeToggle<CR>', A.opts)
-            vim.g.undotree_DiffAutoOpen = false
-            vim.g.undotree_SetFocusWhenToggle = true
-            vim.g.undotree_SplitWidth = 35
-            vim.g.undotree_WindowLayout = 3
-        end
-    },
-    ---- Yanky
-    {
-        'gbprod/yanky.nvim',
-        event = { "BufReadPre", "BufNewFile" },
-        config = function()
-            require("yanky").setup({ highlight = { timer = 100 } })
-            A.map({ "n", "x" }, "y", "<Plug>(YankyYank)", A.opts)
-            A.map({ "n", "x" }, "p", "<Plug>(YankyPutAfter)", A.opts)
-            A.map({ "n", "x" }, "P", "<Plug>(YankyPutBefore)", A.opts)
-            A.map("n", "<C-n>", "<Plug>(YankyNextEntry)", A.opts)
-            A.map("n", "<C-p>", "<Plug>(YankyPreviousEntry)", A.opts)
-        end
-    }
+    require("plugins.autopairs"),
+    require("plugins.comment"),
+    require("plugins.dashboard"),
+    require("plugins.fugitive"),
+    require("plugins.git-signs"),
+    require("plugins.indent-blank-line"),
+    require("plugins.lualine"),
+    require("plugins.markdown-preview"),
+    require("plugins.mason"),
+    require("plugins.lspconfig"),
+    require("plugins.nvim-cmp"),
+    require("plugins.nvim-tree"),
+    require("plugins.surround"),
+    require("plugins.telescope"),
+    require("plugins.tmux"),
+    require("plugins.toggleterm"),
+    require("plugins.tokyonight"),
+    require("plugins.treesitter"),
+    require("plugins.undo-tree"),
+    require("plugins.yanky"),
 }
