+++ lua/plugins/autopairs.lua
@@ -1,0 +1,5 @@
+return {
+    'windwp/nvim-autopairs',
+    event = "InsertEnter",
+    opts = {} -- this is equalent to setup({}) function
+}
