+++ lua/plugins/comment.lua
@@ -1,0 +1,11 @@
+return {
+            'numToStr/Comment.nvim',
+            event = { "BufReadPre", "BufNewFile" },
+            config = function()
+                require('Comment').setup({
+                    toggler = { line = '<Leader>`', block = '<Leader>~', },
+                    opleader = { line = '<Leader>`', block = '<Leader>~', },
+                    mappings = { extra = false, extended = false, },
+                })
+            end
+}
