+++ lua/plugins/surround.lua
@@ -1,0 +1,7 @@
+return {
+        'kylechui/nvim-surround',
+        event = { "BufReadPre", "BufNewFile" },
+        config = function()
+            require("nvim-surround").setup()
+        end
+    }
