+++ lua/plugins/tokyonight.lua
@@ -1,0 +1,13 @@
+return {
+        'folke/tokyonight.nvim',
+        lazy = false,
+        priority = 1000,
+        config = function()
+            require("tokyonight").setup({
+                sidebars = { "qf", "help", "undotree" },
+                lualine_bold = true,
+            })
+            vim.cmd("colorscheme tokyonight")
+        end
+    }
+
