+++ lua/plugins/undo-tree.lua
@@ -1,0 +1,12 @@
+return     {
+        'mbbill/undotree',
+        event = { "BufReadPre", "BufNewFile" },
+        config = function()
+            A.map('n', '<Leader>u', ':UndotreeToggle<CR>', A.opts)
+            vim.g.undotree_DiffAutoOpen = false
+            vim.g.undotree_SetFocusWhenToggle = true
+            vim.g.undotree_SplitWidth = 35
+            vim.g.undotree_WindowLayout = 3
+        end
+    }
+
